Refactor 2
