+++ packages/react-app/src/App.jsx
@@ -30,6 +30,7 @@ import { useBalance, useExchangePrice, useGasPrice, useLocalStorage, usePoller, 
 import WalletConnect from "@walletconnect/client";
 
 import { TransactionManager } from "./helpers/TransactionManager";
+import { EIP618Helper } from "./helpers/EIP618Helper";
 
 const { confirm } = Modal;
 
@@ -922,7 +923,7 @@ function App(props) {
         <div style={{ padding: 10 }}>
           <AddressInput
             selectedChainId={selectedChainId}
-            tx={tx}
+            eip618Helper={new EIP618Helper(tx)}
             ensProvider={mainnetProvider}
             placeholder="to address"
             disabled={walletConnectTx}
